Вынес рендер спрайтов в отдельную директорию
